+++ src/LoginForm.jsx
@@ -1,0 +1,28 @@
+export default function LoginForm() {
+  function handleClick() {
+    // todo
+  }
+
+  return (
+    <div>
+      <div>
+        <label htmlFor="login-email">
+          E-mail
+        </label>
+        <input type="email" id="login-email" />
+      </div>
+      <div>
+        <label htmlFor="login-password">
+          Password
+        </label>
+        <input type="password" id="login-password" />
+      </div>
+      <button
+        type="button"
+        onClick={handleClick}
+      >
+        Log In
+      </button>
+    </div>
+  );
+}
