+++ src/LoginFormContainer.jsx
@@ -2,32 +2,16 @@ import { useDispatch } from 'react-redux';
 
 import { requestLogin } from './actions';
 
+import LoginForm from './LoginForm';
+
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
 
-  function handleClick() {
+  function handleSubmit() {
     dispatch(requestLogin());
   }
+
   return (
-    <div>
-      <div>
-        <label htmlFor="login-email">
-          E-mail
-        </label>
-        <input type="email" id="login-email" />
-      </div>
-      <div>
-        <label htmlFor="login-password">
-          Password
-        </label>
-        <input type="password" id="login-password" />
-      </div>
-      <button
-        type="button"
-        onClick={handleClick}
-      >
-        Log In
-      </button>
-    </div>
+    <LoginForm onSubmit={handleSubmit} />
   );
 }
